pyproject.toml
[project]
name = "demo-pipeline"
version = "0.1.0"
description = "The goal of this repo is to quickly create and evaluate modules in different pipeline setups"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "AI4GOOD Social Media Disinformation Team"}
]
dependencies = [
    "polars==1.35.2",
    "jupyterlab==4.5.0",
    "instaloader==4.15",
    "moviepy==2.2.1",
    "google-genai>=0.3.0",
    "firebase-admin>=6.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.1",
    "google-cloud-firestore>=2.14.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "requests>=2.31.0",
    "librosa>=0.11.0",
    "transformers>=4.57.3",
    "moviepy==1.0.3",
    "torchvision==0.24.1",
    "opencv-python==4.12.0.88"
]

[project.urls]
Repository = "https://github.com/AI4GOOD-social-media-desinformation/demo-pipeline"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "src.dataloaders", "src.modules", "src.pipelines", "src.utils", "src.eventbus", "src.storage"]
